backend-ocr/code/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.2.1
fastapi==0.115.12
Flask==3.1.1
flask-cors==6.0.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
pytesseract==0.3.13
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
Werkzeug==3.1.3
